backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2
eventlet==0.35.1
python-dotenv==1.0.0
werkzeug==3.0.1
sqlalchemy==2.0.25
bcrypt==4.1.2
